"refactor: moved parse functionality to panther directory"
