+++ src/node/panther/parse.changeNodes.ts
@@ -3,10 +3,10 @@ import _ from "lodash"
 import { Unsure } from "../../globals/coding/code.types"
 import { FullPantherEntity, PantherEntity } from "../../globals/panther/models.nodes"
 import { HasConfiguration, HasGeometry, HasInterval, HasLevels, HasUnits } from "../../globals/panther/models.nodes.properties.general"
-import { InvalidRequestError } from "./models.errors"
+import { InvalidRequestError } from "../api/models.errors"
 import { HasBands, HasColor, HasDocumentId, HasSpecificName, HasTimeseries, HasUrl } from "../../globals/panther/models.nodes.properties.datasources"
 import { UsedDatasourceLabels, UsedNodeLabels } from "../../globals/panther/enums.panther"
-import { validateNodeLabels } from "./validations.shared"
+import { validateNodeLabels } from "../api/validations.shared"
 import { isoIntervalToTimestamps, nowTimestamp } from "../../globals/coding/code.dates"
 import { csvParseNumbers, csvParseStrings } from "../../globals/coding/formats.csv"
